utils/youtube_requirements.txt
youtube_transcript_api>=0.6.0
openai>=1.0.0
langdetect>=1.0.9
python-dotenv>=1.0.0 